Add cast to silence warning, cross-dll helpers, improve init code
Added cast to silence a warning when compiling with v140_xp
Added cross-dll helpers for getting cvars and alerting to console
Improved init code, developer cvar pointer now stored as global

+++ engine/eiface.h
@@ -1529,7 +1529,7 @@ typedef enum _fieldtypes
 	FIELD_TYPECOUNT,		// MUST BE LAST
 } FIELDTYPE;
 
-#define _FIELD(type,name,fieldtype,count,flags)				{ fieldtype, #name, offsetof(type, name), count, flags }
+#define _FIELD(type,name,fieldtype,count,flags)				{ fieldtype, #name, static_cast<int>( offsetof(type, name) ), count, flags }
 #define _BASEENT_FIELD( name, fieldtype, count, flags )		_FIELD( ThisClass, name, fieldtype, count, flags )
 #define DEFINE_FIELD( name,fieldtype)						_BASEENT_FIELD( name, fieldtype, 1, 0)
 #define DEFINE_ARRAY( name,fieldtype,count)					_BASEENT_FIELD( name, fieldtype, count, 0)
